fix: rename VercelHelper, remove VercelProvider, fix LDMessage[] input handling
