+++ packages/ai-providers/server-ai-vercel/src/VercelHelper.ts
