+++ packages/ai-providers/server-ai-vercel/src/VercelModelRunner.ts
@@ -9,7 +9,7 @@ import type {
 } from '@launchdarkly/server-sdk-ai';
 
 import type { VercelAIModelParameters } from './types';
-import { convertMessagesToVercel, getAIMetricsFromResponse } from './vercelHelper';
+import { convertMessagesToVercel, getAIMetricsFromResponse } from './VercelHelper';
 
 /**
  * Runner implementation for Vercel AI SDK chat models.
@@ -38,15 +38,17 @@ export class VercelModelRunner implements Runner {
   /**
    * Run the Vercel AI model with the given prompt.
    *
-   * @param input The user prompt to send to the model.
+   * @param input The user prompt string, or a pre-built message array. When a
+   *   string is supplied the config's system messages are prepended automatically.
+   *   When a `LDMessage[]` is supplied it is used as-is (config messages are NOT
+   *   prepended — the caller is responsible for the full message list).
    * @param outputType Optional JSON schema for structured output. When provided,
    *   the parsed object is exposed via {@link RunnerResult.parsed}.
    */
-  async run(input: string, outputType?: Record<string, unknown>): Promise<RunnerResult> {
-    const messages: LDMessage[] = [
-      ...(this._config.messages ?? []),
-      { role: 'user', content: input },
-    ];
+  async run(input: string | LDMessage[], outputType?: Record<string, unknown>): Promise<RunnerResult> {
+    const messages: LDMessage[] = Array.isArray(input)
+      ? input
+      : [...(this._config.messages ?? []), { role: 'user', content: input }];
 
     if (outputType !== undefined) {
       return this._runStructured(messages, outputType);
